(chore) rename component to components.
