+++ src/app/app.module.ts
@@ -7,7 +7,7 @@ import { IgoModule, LanguageLoader, provideLanguageService,
 import { AppComponent } from './app.component';
 import { CoreModule } from './core/core.module';
 import { SharedModule } from './shared/shared.module';
-import { ComponentModule } from './component/component.module';
+import { ComponentModule } from './components/components.module';
 import { InterventionReportModule } from './intervention-report/intervention-report.module';
 import { InterventionSurveyModule } from './intervention-survey/intervention-survey.module';
 import { PreventionSurveyModule } from './prevention-survey/prevention-survey.module';
